Partially fix items not going away
Due to how the plugin reads the config, it falls back onto the default value if the value is not present. To get around this issue, I made slot #0 the slot that can't be removed. This can reduce issues due to it being a single slot and can simply have it's material set to 'air'.